Add attacks to score function
It might make sense to remove this later, or at least restrict it to only the final search, since this just approximates the score of the position one half turn later. I think something like this is necessary to avoid blunders though -- maybe we just need to consider the enemy's attacks on us though?